refactor: move posts code under /src and extract posts content
