--- posts/consts.ts
@@ -1,5 +1,0 @@
-const postsBasePath = 'posts/db';
-
-export {
-    postsBasePath
-}
