--- posts/db/hello-world.md
@@ -1,40 +1,0 @@
----
-id: hello-world
-title: First post
-createdDate: 2020-02-19
----
-
-# Hello world
-Super awesome post.
-
-
-What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the printing
-and typesetting industry. Lorem Ipsum has been the industry's standard
-dummy text ever since the 1500s, when an unknown printer took a galley
-of type and scrambled it to make a type specimen book. It has survived
-not only five centuries, but also the leap into electronic typesetting,
-remaining essentially unchanged. It was popularised in the 1960s with
-the release of Letraset sheets containing Lorem Ipsum passages, and more
-recently with desktop publishing software like Aldus PageMaker including
-versions of Lorem Ipsum.
-
-## Subtitle 2
-
-What would I like to write about here? I have no idea yeat.
-This is some test text to fine tune styling and see how it will look like...
-Test test test Lorem Ipsum testy test.
-
-### Subtitle 3
-Check this link: https://onet.pl
-
-
-Why do we use it? It is a long established fact that a reader will be
-distracted by the readable content of a page when looking at its layout.
-The point of using Lorem Ipsum is that it has a more-or-less normal
-distribution of letters, as opposed to using 'Content here, content
-here', making it look like readable English. Many desktop publishing
-packages and web page editors now use Lorem Ipsum as their default model
-text, and a search for 'lorem ipsum' will uncover many web sites still
-in their infancy. Various versions have evolved over the years,
-sometimes by accident, sometimes on purpose (injected humour and the
-like).
