+++ src/consts.ts
@@ -1,0 +1,5 @@
+const postsBasePath = 'db';
+
+export {
+    postsBasePath
+}
