+++ src/posts/getPost.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import { promisify } from 'util';
-import { postsBasePath } from "./consts";
+import { postsBasePath } from "../consts";
 
 import { convertMarkdownToPost } from './convertMarkdownToPost';
